StudArt/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
